Exposed methods to get data from url
There are now API calls available to grab from an arbitrary url.  Of
course, when using this, the user must know what data he/she is
expecting from the call.  Thus, there is one for getting a single card,
a single set, a group of cards, and a group of sets (soon).